Build!: Adds Type assertion for _dog in dates.ts.
Details: Something is wrong with the DateCalculator class. I am not sure
if it the _dog object or some kind of typeguard. Either way, this change
SHOULD be applicable eventually. I'm implementing it now because I'm not
sure that typescript is evaluating all of the type guards for it
correctly.